requirements.txt
python-telegram-bot[job-queue]>=20.0
yt-dlp>=2023.10.13
pysubs2>=1.6.0
aiofiles>=23.2.0
aiohttp>=3.9.0
python-dotenv>=1.0.0